remove old jsonapi-serializer internal typings file
